InputContainer test 구현
Page component에 있는 Input과 관련된 내용을 InputContainer로 옮겼습니다.
